# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IKODI Student Sponsorship Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools:
    get:
      operationId: listSchools
      tags: [schools]
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [primary, secondary, higher_learning]
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      operationId: createSchool
      tags: [schools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolBody"
      responses:
        "201":
          description: School created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /schools/{id}:
    get:
      operationId: getSchool
      tags: [schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: School details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    put:
      operationId: updateSchool
      tags: [schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolBody"
      responses:
        "200":
          description: School updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
...
  /sponsors:
    get:
      operationId: listSponsors
      tags: [sponsors]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [individual, organization]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: List of sponsors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sponsor"
    post:
      operationId: createSponsor
      tags: [sponsors]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSponsorBody"
      responses:
        "201":
          description: Sponsor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
  /sponsors/{id}:
    get:
      operationId: getSponsor
      tags: [sponsors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sponsor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorDetail"
    put:
      operationId: updateSponsor
      tags: [sponsors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSponsorBody"
      responses:
        "200":
          description: Sponsor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsor"
    delete:
      operationId: deleteSponsor
      tags: [sponsors]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Sponsor deleted
  /sponsorships:
    get:
      operationId: listSponsorships
      tags: [sponsorships]
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: sponsorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, completed]
        - name: term
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sponsorships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sponsorship"
    post:
      operationId: createSponsorship
      tags: [sponsorships]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSponsorshipBody"
      responses:
        "201":
          description: Sponsorship created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsorship"
  /sponsorships/{id}:
    get:
      operationId: getSponsorship
      tags: [sponsorships]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sponsorship details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsorship"
    put:
      operationId: updateSponsorship
      tags: [sponsorships]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSponsorshipBody"
      responses:
        "200":
          description: Sponsorship updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sponsorship"
    delete:
      operationId: deleteSponsorship
      tags: [sponsorships]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Sponsorship deleted
...
components:
  schemas:
...
    School:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [primary, secondary, higher_learning]
        location:
          type: string
        contactPhone:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - category
        - location
        - createdAt
    CreateSchoolBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [primary, secondary, higher_learning]
        location:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
      required:
        - name
        - category
        - location
...
    Sponsor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [individual, organization]
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        totalContributed:
          type: number
        activeStudents:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - status
        - totalContributed
        - activeStudents
        - createdAt
    SponsorDetail:
      allOf:
        - $ref: "#/components/schemas/Sponsor"
        - type: object
          properties:
            sponsorships:
              type: array
              items:
                $ref: "#/components/schemas/Sponsorship"
            recentPayments:
              type: array
              items:
                $ref: "#/components/schemas/Payment"
    CreateSponsorBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [individual, organization]
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
      required:
        - name
        - type
    Sponsorship:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        sponsorId:
          type: integer
        studentName:
          type: string
          nullable: true
        sponsorName:
          type: string
          nullable: true
        coverageType:
          type: string
          enum: [full, partial]
        amount:
          type: number
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, completed]
        term:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - studentId
        - sponsorId
        - coverageType
        - amount
        - startDate
        - status
        - createdAt
    CreateSponsorshipBody:
      type: object
      properties:
        studentId:
          type: integer
        sponsorId:
          type: integer
        coverageType:
          type: string
          enum: [full, partial]
        amount:
          type: number
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        term:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, completed]
      required:
        - studentId
        - sponsorId
        - coverageType
        - amount
        - startDate
    UpdateSponsorshipBody:
      type: object
      properties:
        coverageType:
          type: string
          enum: [full, partial]
        amount:
          type: number
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, completed]
        notes:
          type: string
          nullable: true
    Payment:
      type: object
      properties:
        id:
          type: integer
        sponsorshipId:
          type: integer
          nullable: true
        sponsorId:
          type: integer
          nullable: true
        studentId:
          type: integer
          nullable: true
        studentName:
          type: string
          nullable: true
        sponsorName:
          type: string
          nullable: true
        amount:
          type: number
        paymentDate:
          type: string
        paymentMethod:
          type: string
          enum: [mpesa, bank_transfer, cash, cheque, online]
        referenceNumber:
          type: string
          nullable: true
        purpose:
          type: string
          nullable: true
        term:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - amount
        - paymentDate
        - paymentMethod
        - createdAt
...